Add the ability to set the expected true/false
You can set a default for the whole validate,
and you can set an expected value for each
validator.

This makes it easier to use functions that
already return a boolean in Elixir such as
String.contains?/2. Now you can more easily
validate the presence or absence of something
with that function, for example.

co-authored-by: Sukaant Chaudhary <sukaant.chaudhary@gmail.com>